[docs] Remove body from minimizing output bits recipe (#2290)
Remove the body from the emitted Verilog. This was the original intent
of the example, and it avoids an issue where Jekyll was not able to
render the Markdown file due to Verilog concatenation looking like a
variable escape.

(cherry picked from commit e85bfebb5d661de41f9ccac300fb48bf92840cfe)