Fix warning injected into user code by @chiselName (#2500)
In Scala 2.13, Auto-application to `()` is deprecated. Any nullary
method (ie. a method with no arguments) that is defined with () must now
be called with (). @chiselName used to inject a case of this warning
into user code which would then cause a warning on the @chiselName macro
that the user has no way to fix.

(cherry picked from commit ca902680df772445371e6c3dd907c01113afb1f5)