Remove "don't touch" behavior from Trace API
Change the behavior of the Trace API to not have "don't touch" behavior
unconditionally.  Removing "don't touch" enables a finer granularity of
functionality: if a user wants "don't touch" they can add a "don't
touch".  However, they are not locked optimization blocking behavior
which may be undesirable in certain situations.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>