WIP on trying to save a contains check in Namespace
It relies on the observation that when a name is taken and we have
determined our disambiguation index, it's possible we don't actually
need to check the HashMap again to see if name + index is in it. We
would only need to do so if previously there has been a name added that
shares our prefix and happens to have an index itself. This commit is a
WIP to make this work, but it's quite hard to get right and the normal
names.get path needs to also do a prefix check. See the final test in
NamespaceSpec which fails (yay good tests!)