Optimize _ids (#2866)
* Don't copy _ids to a Seq just to Iterate on them

Saves a copy of a large data structure.

* Only add Data to _ids when the binding is constrained

ConstrainedBinding means that the bound Data is scoped to a module.
Since _ids are used for naming of signals within a module, there is no
reason to add Unconstrained Data to the _ids of the Module they were
constructed during. Unconstrained Data include things like literals and
children of Aggregates (which sort of are constrained to a Module but
only via their parent, they are not directly named by the module).