[circt] Cleanup CIRCT error behavior
Change the behavior of what happens if firtool returns a non-zero exit
code from printing information to stdout to throwing an exception with
this information in the body.  This is done primarily to ease
interoperability with mdoc.  However, this is generally better
principled than the previous behavior of scattering information in
different mediums.  Add a test of this behavior.

Stop intercepting a file not found error if CIRCT fails to print this to
stdout and wrap this in a StageError.  It is better to just let the file
not found exception through as this already includes all the information
necessary to tell the user what the problem is.  Do not test this
behavior (as it is tedious to test).

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>