rewrite mill build scripts for chisel5.
This PR cleaned up the old build script, split it to common.sc and
builds.sc two files. common.sc is used for declare the dependencies and
project structual for each modules, which can be reused by other
projects for building from source. build.sc is used for build project in
tree.
This commit makes chisel3['2.13.10'] being able to compile.