Fix calculuation of firrtl directions for Vecs
For ports of type Vec, the direction of both the Vec itself and the
direction of its elements can affect the direction of the FIRRTL port.
Previously, we did not handle the fact that the Vec's elements could,
themselves, be Vecs, thus making this directionality calculation
recursive. Also, the logic was only applying to regular RawModules and
Modules. The fix now applies to BlackBoxes and ExtModules as well.

(cherry picked from commit b501b6994683151dd3752efa4ad4ffac432659fc)

# Conflicts:
#	core/src/main/scala/chisel3/RawModule.scala
#	src/test/scala/chiselTests/Direction.scala