Convert loadMemoryFromFile tests to use MFC
Convert/enable testing of loadMemoryFromFile and
loadMemoryFromFileInline to use the MLIR-based FIRRTL Compiler (MFC).
The only change to the tests is that MFC will _not_ lower aggregate
memories.  (This is an intentional behavior that is a full realization
of the "SimplifyMems" pass in SFC, but extended to work with mask bits.)

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>