[CIRCT] Remove wrong ReplSeqMemAnno to bbox arg
Remove an incorrect code path in the CIRCT phase that would convert a
"ReplSeqMemAnnotation" into a "-blackbox" argument.  This option has
existed for a long time and this would just cause "firtool" to error.
Users will need to use the dedicated "-repl-seq-mem" options to
"firtool" going forward (which is exactly what they have to do now).

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>