Convert tests to circt.stage.ChiselStage
Modify tests, other than ChiselStageSpec, to all use
circt.stage.ChiselStage and not chisel3.stage.ChiselStage.  This paves
the way for the deletion of the latter in Chisel 5.

Note: there is one bug in the Trace API resulting in a commented out
test.  This is tracked with this CIRCT issue:

  - https://github.com/llvm/circt/issues/4661

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>