Pass some annotations around firtool (#3015)
Make ImportDefinitionAnnotation and DesignAnnotation pass around the
firtool call in the CIRCT phase so that these can be used by later
Chisel elaborations which may chain off of an initial ChiselStage call.
(This is a frequent pattern in D/I and is used in
SeparableElaborationSpec.)

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>

Make this passthrough along with DesignAnnotation.