Deprecate CIRCTStage (firtool wrapper)
Deprecate CIRCTStage and CIRCTMain.  These are stage-based wrappers
around just running firtool that are roughly equivalent to FirrtlStage
and FirrtlMain that did this for the SFC.  These do not have users and
it is unclear why these should persist given that
circt.stage.ChiselStage subsumes almost all use cases (going Chisel to
Verilog or Chisel to CHIRRTL and then using the build system to go from
CHIRRTL to Verilog).

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>