[chisel5] Remove all WIR that is not IR
Remove everything from WIR.scala (Working IR) which is not actually IR.
I.e., this removes everything except CHIRRTL constructs and Kind/Flow.
A method is moved into CheckTypes that is used by Chisel to determine
connection validity.  This was unnecessarily bouncing through
WrappedType when it can be simplified to not do the boxing.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>