Add "unipublish" for single artifact publishing
It is a new pseduo-project in the SBT build that aggregates build
artifacts from the real projects into a single artifact. It must now be
used for publishing and MiMa binary compatibility checking.

There are also other changes needed for publishing under
org.chipsalliance, namely we have to publish to a different sonatype
snapshots repository: https://s01.oss.sonatype.org/