Use -std=c++11 for Verilator 5.0 support (#3066)
Verilator 5.0 emits C++ which requires "-std=c++11".  This is normally
handled by the Verilator configure script.  However, since Chisel
doesn't use this, the lack of this when running with Verilator 5.0 can
result in compilation failures in tests.  Add this flag so that
Verilator 5.0 works.  I have not tested this with earlier Verilator
versions.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>