Use %[[]] format for ChsielAnnotations (#3141)
Change FIRRTL Circuit conversion/serialization to place
ChiselAnnotations (converted to FIRRTL Annotations) inside the FIRRTL
Circuit and to serialize this using the "%[[]]" inline JSON format that
is present in the FIRRTL spec.  This does not remove _all_ annotations
passed via the annotation file, but it is the first step towards that.