[CI] Add Release Notes Automation (#3170)
Add new workflow to generate release notes. This workflow can be run
manually via workflow dispatch, but will also run automatically upon
creating a release. The workflow dispatch can be used for testing,
previewing release notes for a potential release, or for generating
release notes for past releases for when the automation has a bug or
didn't exist.

Also update the PR template and CONTRIBUTING guidelines to include
information about the new required labels and automation. Also
streamline the PR template a little bit to nudge developers towards
putting more information in the "Release Notes" section of the PR
template.