Emit annotations in the .fir file
This commit reimplements the previous work adding emission of
annotations via the FIRRTL specification v2.0.0 %[[...]] style.
Now all annotations that were previously emitted to the .anno.json file
are emitted in the .fir file. This uses a slightly hacky new internal
mechanism in firrtl that is already marked deprecated to discourage
users from touching it.