Add --module-json option to ChiselStage
Add a new option, "--module-json", to construct a
ChiselGeneratorAnnotation, from a class name and some JSON text.  The
JSON text, using "class" full type hints, will be deserialized to an
object and passed as the single parameter to a Chisel generator.  This
is intended to be used by tools which want to do "inline generation" (to
call out to a Chisel generator to fill in a module body, e.g., by
Firesim).

This is available by default by ChiselStage and can be used by any
non-ChiselStage tool that supports ChiselGeneratorAnnotation.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>