[interface] Fix Wrapper.Stub Name (#3225)
Fix a bug where the name of the Stub module (a module auto-generated from
an Interface without an implementation available) would be called "Stub"
as opposed to the name of the Interface.  This could create a linker error
if the Stub was used.

Fix this setting the final "desiredName" inside the Entity.  Add a test that
this works for the Stub.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>