[bugfix] Fix BoringUtils up--down bug
Fix a bug in the new BoringUtils.bore API where this would generate
invalid FIRRTL if the user did boring which created a path that went up to
an LCA and then down.  Change the internal leastCommonAncestor utility to
also return the LCA module as this avoids a second call to compute the
LCA.

Fixes #3239.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>