[bugfix] Don't return Data from BoringUtils APIs
Fix a bug in BoringUtils (and secret IO APIs) where APIs would return a
Data.  This could create problems if the BoringUtils APIs were used
directly in operations, e.g., the following could cause a problem:

    val a = b & BoringUtils.bore(c)

This requires some casting of DontCare as it is unfortunately not a proper
"bottom" type.

No test is added as the test is that Scala passes type checking.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>