Fix issue where reify would do the same lookup twice (#3392)
This is not a functional bug. The recursive form of reify for Element
would use the .topBinding from the wrong argument, so it would recurse
an extra time, but it would ultimately converge to the right value.

(cherry picked from commit 374b45b9058ecd418955d06621e98fe3bc1e67bb)