[ci] Scala-CLI Template Action -> Workflow, NFC
Change the existing Scala-CLI composite action to a workflow.  Update
other workflows which now need to work a little differently.  The main
change is that a reusable workflow can only communicate through
uploads/downloads.  Hence, change the job that publishes the Scala-CLI
template to download this from the new, reusable workflow.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>