Switch to firtool-resolver
Whereas in the past, the user was expected to provide the firtool binary
on the PATH, Chisel will now manage the firtool binary automatically.
Chisel users the following 3-step process for determining which binary
to use:

1. If environment variable FIRTOOL_PATH is set, it will check for
   firtool in that directory
2. It will check the classpath for firtool provided by
   chipsalliance/llvm-firtool
3. It will download a default version of chipsalliance/llvm-firtool

Both (2) and (3) will store the firtool binary in a sensible location
depending on the user's operating system. See dirs-dev/directories-jvm
for more information. The user can override this location by setting the
FIRTOOL_CACHE environment variable.