[ci] Change Scala-CLI Template Composite Action to Workflow (#3457)
* [ci] Scala-CLI Template Action -> Workflow, NFC

Change the existing Scala-CLI composite action to a workflow.  Update
other workflows which now need to work a little differently.  The main
change is that a reusable workflow can only communicate through
uploads/downloads.  Hence, change the job that publishes the Scala-CLI
template to download this from the new, reusable workflow.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>

* [ci] Display Scala-CLI Template in Summary, NFC

Show the resulting Scala-CLI template as a "GITHUB_STEP_SUMMARY" in the
GitHub action for publishing it.  This is likely more useful than making
it available as a download artifact.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>

* [ci] Whitespace cleanup, NFC

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>

---------

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>