[ci] Use 'default' as version string for typical CIRCT version
Make it such that using "Install CIRCT" does not require specifying a
CIRCT version, instead using the string 'default' as a way to pick the
"default version". The "default version" is the version that Chisel is
tested and released against.

This fixes "Generate Scala CLI Template" which does not care to specify
a CIRCT version.