[ci] Remove caching of firtool
Remove caching of the `firtool` binary.  Caching empirically provides no
measurable benefit over downloading the binary from a GitHub Release every
time.

This has the effect of avoiding a bug where the "default" old version
would incorrectly hit when the `firtool` version is updated via a change
to the `./etc/circt.json` file.

Additionally, fix another bug where the environment is not populated for
composite actions---the `./etc/circt.json` filename needs to be inlined to
work.  This was previously working because the cache was always hitting.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>