Make LTL properties use Clock and Disable by default (#3498)
Properties in package chisel3.ltl will now be clocked and disabled by
default (if there is an implicit clock and disable). Default disable is
not hasBeenReset of the current implicit reset (if one exists). The
clock and disable can be removed by setting them to None via withClock,
withReset, and withDisable APIs.