Add ability to block IO creation
Add a `private[chisel3]` ability to block further IO creation in a
`BaseModule`.  This is intended for advanced, in-tree APIs that need to
disallow the creation of further IO by a user.  This API has mutable state
to track if more IO is allowed to be created, a method to get the
value, and a method to set the module to disallow further IO creation.
This one-way setting (h/t @jackkoenig) tries to keep this as locked down
as possible.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>