Add FixedIOModule, FixedIOBlackBox
Add a new module hierarchy that promotes the type of the IO of that module
to a type parameter.  This module hierarchy includes:

  1. FixedIOBaseModule
  2. FixedIORawModule
  3. FixedIOBlackBox

Both (2) and (3) inherit from (1).  These modules are all "fixed" in the
type of their IO.  It follows that no more IO are allowed to be created
other than what is in the type.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>