[ci] Bump install-circt to v1.1.1 (#3586)
Switch from install-circt v1.0.0 to v1.1.1.  There is no major change to
functionality other than this switches to using a shell script instead of
scripting in a GitHub composite action.  This also fixes a bug related to the
action not working in a container environment.  (This is not a problem for
Chisel.)

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>