Rename "groups" to "layers"
Rename the "groups" API to "layers".  This is done to match a name change
to the FIRRTL feature (of the same name).  "Layers" better captures the
intended use case.  Namely, a layer is a collection of functionality that
is included as a unit.  Layers are made up of "Layer Blocks" that exist in
one or more modules.

This does not change any FIRRTL codegen.  That will have to happen once
these changes are in CIRCT.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>