Make Property an Element.
This is the right place in the Chisel type hierarchy for Property as
it exists today. Property types can be used in Aggregates, and are not
themselves Aggregates in their current form. Making Property extend
Element allows much of Chisel's internal reflection to work as
expected when Property types are used in Aggregates.