Exclude docs project from Unidoc
The docs project includes logic only used for generating mdoc, it should
not be included in the generated ScalaDoc as it is not actually included
in the published Chisel artifacts.

This also helps avoid flaky CI issues--the publish step will no longer
compile the docs subproject. Compiling the docs subproject requires
scala-cli which requires Java 17. Downloading Java 17 at build time can
timeout, causing flaky CI. Note that the publish step uses Java 8 so
that we can guarantee the published artifacts work on Java 8.