[CI] Make SBT version handling logic more robust
The automation around automatically determining released versions of
Chisel and generating content for the website would fail in the period
between tagging a new release and the corresponding artifacts actually
being released to Maven Central. This would even prevent the CI job
trying to push the release to Maven Central.

Now, the logic for determining releases will filter for those actually
pushed to Maven Central.