Add a new DynamicObject.apply method to create Class instances. (#3792)
This is in addition to the existing support in
Definition.apply. Sometimes it is not possible to use
Definition.apply, for example, if you plan to bore ports through the
Class being constructed. The new DynamicObject.apply method supports
this, and creates a DynamicObject from the newly elaborated Class.

(cherry picked from commit 04535d325984cc83a7bbf81d55371f4212e06ae1)