Bump to FIRRTL 4.0.0
Bump emission to use FIRRTL 4.0.0.  Since Chisel is now emitting layer-colored
probes and the work-in-progress "enablelayer" keyword, Chisel needs to emit a
version of FIRRTL that compilers will accept.

The only change to the output (other than the stated version) is "groups" are
now changed to "layers".

Note: FIRRTL 4.0.0 is not finalized yet.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>