Add test of external modules to ChiselSim tests
This tests that compilation with an external module works (as well as
testing that reading from a port driven from an external module
works---which doesn't provide any additional coverage).  This is
highlighting a problem in CIRCT beyond 1.66.0 where this relies on a
`filelist.f` file being produced which includes information about all
external modules.  This was dropped in a later commit [[1]] and needs to
be fixed.  This commit is intending to cause nightly to fail, demonstrate
the problem, and provide a check so that we know things are back to
working.

[1]: https://github.com/llvm/circt/pull/6729

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>