Get ChiselSim working with CIRCT 1.66+
Fix an incorrect assumption in ChiselSim that a 'filelist.f' will exist
that includes all modules, _including blackboxes_, in the design.  Change
this to try to also copy files from the black box filelist.  This makes
ChiselSim work with CIRCT 1.66 and earlier (which put everything in
'filelist.f') and CIRCT 1.66+ (which does not include blackboxes in
'filelist.f').

Note that this is intended as a temporary solution until FIRRTL has a rock
solid ABI that is implemented by CIRCT where all modules can be discovered
under an arbitrary public module.

See CIRCT changes here: https://github.com/llvm/circt/pull/6729

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>