[chiselsim] Better fix for CIRCT 1.66- and 1.66+ (#3893)
Change the way ChiselSim moves files to track and not move files which may
be duplicated in both the default black box filelist and in the main
CIRCT-produced filelist.  Previously, this was working because a java.nio
exception was being ignored.  However, this may have been hiding bugs
where the filelist was actually pointing at a file that did not exist.

This commit also does some heavy cleanup of the code for moving files to only
use java.nio methods as opposed to relying on doing string manipulations of
paths.  The logic is intended to be the exact same as before.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>
(cherry picked from commit 61df8bbf6a4dc097a0ef2a21051c13d1e70edded)