Allow BoringUtils.bore to work on probes (#3908)
Extends BorginUtils.bore (and related) to work where the source is a
probe.  Currently, if a user tries to do this, they will get an error
about being unable to probe a probe because the old code would
unconditionally wrap whatever the source was in `Probe`/`RWProbe`.  Change
this to not wrap if the source is already a probe.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>
(cherry picked from commit 67f51b884e841cd088f8aff47e9aa003f7628563)