Add a label "chisel3_builtin" to ifElseFatal intrinsic
This is a follow-up to #3825.

MFC has attached a laebl "chisel3_builtin" to every ifElseFatal assertion.
Usually a label on ifElseFatal is just ignored but the label is emitted
when compiled with `-emit-chisel-asserts-as-sva`. So labels on assertions
converted from ifElseFatal disappear due to #3825. This PR restores labels
to preserve the old behavior.